refactoring back end
