+++ Jesus Flores/week_11/ba-express-mongodb-backend/server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const app = express();
 const PORT = 3000;
 const db = require('./dbconfig');
-const Flight = require('./models/Flight');
+const flightController = require('./controllers/flightsController');
 const cors = require('cors')
 app.use(cors())
 // Parse URL-encoded bodies (as sent by HTML forms)
@@ -18,42 +18,10 @@ app.listen(PORT, () => {
     console.log(`server running on port ${PORT}`)
 })
 
-app.get('/', (req, res) => {
-    console.log('Welcome to BA Backend');
-    res.send('Welcome to BA Backend');
-})
+app.get('/', flightController.home )
 
-app.get('/flights/search/:origin/:destination', async (req, res) => {
-    try {
-        const {origin , destination } = req.params
-        const flights = await Flight.find({ origin, destination});
-        res.json(flights)
-    } catch (error) {
-        console.log(error)
-        res.sendStatus(422); //Unprocesable Entity
-    }
-})
+app.get('/flights/search/:origin/:destination', flightController.searchFlight)
 
-app.get('/flights/:id', async (req, res) => {
-    console.log('all good');
-    const flight = await Flight.findOne({_id: req.params.id})
-    res.json(flight)
-})
+app.get('/flights/:id', flightController.getFlight)
 
-app.post('/reservation', async (req, res) => {
-    const newReservation = {
-        row: req.body.row,
-        col: req.body.col,
-        user_id: 11
-    }
-    const flight = await Flight.updateOne(
-        { _id: req.body.flight_id },
-        {
-            $push: {
-                reservations: newReservation
-            }
-        }
-
-    )
-    res.json(newReservation);
-})
+app.post('/reservation', flightController.bookFlight)
